+++ Frontend/app/main/components/contacts/contacts.component.ts
@@ -11,20 +11,26 @@ import {ToastsManager} from 'ng2-toastr';
 @Component({
 	selector:'contacts-component',
 	template: `
-		<button class="btn btn-block" [routerLink]="['/Add-Contact']">Add A Contact</button>
-		<table class="table table-bordered table-justified table-hover">
-			<tbody>
-				<tr class="crm-list-item" *ngFor="let contact of (contacts$ | async)">
-					<td class="crm-list-item-title" (click)="routeWithDispatch(contact, ['/Contact-Details'])">{{contact.name}}</td>
-					<td>
-						<span class="icon icon-share"(click)="routeWithDispatch(contact, ['/Quotes'])"></span>
-					</td>
-				</tr>
-			</tbody>
-		</table>
+		<div *ngIf="!dataReady">
+			<data-loading-screen></data-loading-screen>
+		</div>
+		<div *ngIf="!!dataReady">
+			<button class="btn btn-block" [routerLink]="['/Add-Contact']">Add A Contact</button>
+			<table class="table table-bordered table-justified table-hover">
+				<tbody>
+					<tr class="crm-list-item" *ngFor="let contact of (contacts$ | async)">
+						<td class="crm-list-item-title" (click)="routeWithDispatch(contact, ['/Contact-Details'])">{{contact.name}}</td>
+						<td>
+							<span class="icon icon-share"(click)="routeWithDispatch(contact, ['/Quotes'])"></span>
+						</td>
+					</tr>
+				</tbody>
+			</table>
+		</div>
 	`
 })
 export class ContactsComponent implements OnInit, OnDestroy{
+	public dataReady: boolean = false;
 	private contactsSource: BehaviorSubject<Contact[]> = new BehaviorSubject<Contact[]>([]);
 	public contacts$: Observable<Contact[]> = this.contactsSource.asObservable();
 	public stateSub: Subscription;
@@ -51,11 +57,13 @@ export class ContactsComponent implements OnInit, OnDestroy{
 				this.crmData.getContacts({owner_id: state.selectedCompany.id})
 					.then((contacts: Contact[]) => {
 						this.contactsSource.next(contacts);
+						this.dataReady = true;
 					})
 			} else {
 				this.crmData.getContacts({})
 					.then((contacts: Contact[])=> {
-						this.contactsSource.next(contacts)
+						this.contactsSource.next(contacts);
+						this.dataReady = true;
 					});
 			}
 		});