+++ Frontend/app/main/components/quotes/quotes.component.ts
@@ -14,26 +14,32 @@ import {CRMType} from '../../models/crm-models.type';
 @Component({
 	selector: 'quotes-component',
 	template: `
-		<button class="btn btn-block" [routerLink]="['/Add-Quote']">Create A Quote</button>
-		<ul class="crm-list">
-			<li class="crm-list-item" *ngFor="let quote of (quotes$ | async)">
-				<quote-icons class="crm-list-item-icons">
-					<button class="btn btn-sm btn-default dropdown-toggle" type="button" id="contactDropDown" data-toggle="dropdown" aria-haspopup="true" aria-expanded="true">
-						<span class="icon icon-share"></span>
-					</button>
-					<ul class="dropdown-menu contact" aria-labelledby="contactDropDown">
-						Share With:
-						<li *ngFor="let contact of contactsWithCompanyID(quote.company_id)" (click)="routeWithContactDispatch(contact, quote, ['/Quote-Template'])"><strong>{{contact.name}}</strong></li>
-					</ul>
-					
-				</quote-icons>
-				<quote-name class="crm-list-item-title" (click)="routeWithDispatch(quote, ['/Quote'])">{{quote.name}}
-				</quote-name>
-			</li>
-		</ul>
+		<div *ngIf="!dataReady">
+			<data-loading-screen></data-loading-screen>
+		</div>
+		<div *ngIf="!!dataReady">
+			<button class="btn btn-block" [routerLink]="['/Add-Quote']">Create A Quote</button>
+			<ul class="crm-list">
+				<li class="crm-list-item" *ngFor="let quote of (quotes$ | async)">
+					<quote-icons class="crm-list-item-icons">
+						<button class="btn btn-sm btn-default dropdown-toggle" type="button" id="contactDropDown" data-toggle="dropdown" aria-haspopup="true" aria-expanded="true">
+							<span class="icon icon-share"></span>
+						</button>
+						<ul class="dropdown-menu contact" aria-labelledby="contactDropDown">
+							Share With:
+							<li *ngFor="let contact of contactsWithCompanyID(quote.company_id)" (click)="routeWithContactDispatch(contact, quote, ['/Quote-Template'])"><strong>{{contact.name}}</strong></li>
+						</ul>
+						
+					</quote-icons>
+					<quote-name class="crm-list-item-title" (click)="routeWithDispatch(quote, ['/Quote'])">{{quote.name}}
+					</quote-name>
+				</li>
+			</ul>
+		</div>
 	`
 })
 export class QuotesComponent implements OnInit, OnDestroy {
+	public dataReady: boolean = false;
 	private quotesSource: BehaviorSubject<Quote[]> = new BehaviorSubject<Quote[]>([]);
 	public quotes$: Observable<Quote[]> = this.quotesSource.asObservable();
 	private contactsSource: BehaviorSubject<Contact[]> = new BehaviorSubject<Contact[]>([]);
@@ -56,11 +62,13 @@ export class QuotesComponent implements OnInit, OnDestroy {
 				this.crmData.getQuotes({owner_id: state.selectedCompany.id})
 					.then((quotes: Quote[] )=> {
 						this.quotesSource.next(quotes);
+						this.dataReady = true;
 				})
 			} else {
 				this.crmData.getQuotes({})
 					.then((quotes: Quote[]) => {
 						this.quotesSource.next(quotes);
+						this.dataReady = true;
 				});
 			}
 		});